Create Hacked research
Discovered some low vulnerability going on due to the security pendation. All hackers are expected to be careful/using a vpn to explore  and it's protected by wild Firefox so we suggest reading and studying the HTML.

An attacker can inject HTML on the page and potentially run attacks involving user interaction, with achieving arbitrary javascript code execution not being possible due to the Content Security Policy installed on the server.